CH-303: fix: initialise custom and taxonomies array elements separately - move custom field const to ServiceEligiblity class
