+++ app/Models/ServiceEligibility.php
@@ -8,6 +8,18 @@ class ServiceEligibility extends Model
 {
     use ServiceEligibilityRelationships;
 
+    const SUPPORTED_CUSTOM_FIELD_NAMES = [
+        'age_group',
+        'disability',
+        'employment',
+        'gender',
+        'housing',
+        'income',
+        'language',
+        'ethnicity',
+        'other',
+    ];
+
     /**
      * @return \App\Models\ServiceTaxonomy
      */
